+++ src/main/java/dev/waamir/hotelaluraapi/domain/port/IUserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface IUserRepository<T extends User> {
     T create(T t);
-    Optional<T> getById(int id);
+    Optional<T> getById(Long id);
     Optional<T> getByUsername(String username);
     void delete(T t);
     void update(T t);
